+++ test/support/durable_term_cold_fixture.ex
@@ -1,0 +1,5 @@
+defmodule Continuum.Test.DurableTermColdFixture do
+  @moduledoc false
+
+  def value, do: :continuum_durable_term_cold_fixture_atom_v080
+end